+++ config-http-server/src/main/java/io/scalecube/config/http/server/ConfigRegistryResource.java
@@ -8,13 +8,15 @@ import io.scalecube.config.source.ConfigSourceInfo;
 
 import java.util.Collection;
 
+import javax.annotation.security.PermitAll;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-@Path("/configuration/")
+@Path("/_config/")
 @Produces(MediaType.APPLICATION_JSON)
+@PermitAll // TODO: Require basic auth here instead (see @RolesAllowed)
 public class ConfigRegistryResource {
   private final ConfigRegistry configRegistry;