+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## 0.1.1
+
+* Change HTTP resource base path to `_config`
+* Add shortcut methods to get property values from `ConfigRegistry`
+* Add `addFisrtSource` and `addBeforeSource` methods to `ConfigRegistrySettings`
+
 ## 0.1.0 / 2017-08-15
 
 * Initial development release